gh#27080 Move focus tests into their own spec file
The four advanced-edit focus tests added in 11689fa307c (ALT+E focus
placement, Tab stays in modal, typing doesn't steal focus, close+reopen)
are not really about keyboard shortcuts — only the ALT+E trigger is a
shortcut; the assertions are about DOM focus traversal and React
re-render stability. Per e2e/CLAUDE.md's "group by workflow" rule, they
belong in a dedicated spec, not alongside the shortcut-registration
regression tests from me03#27082.

- New file: e2e/frontend-webui/tests/spec/advanced-edit-focus.spec.js
  with 5 tests (the 4 moved ones + a new one that runs the focus
  assertion against Bank Statement window 194, the exact window from
  the original bug report).
- Reverted e2e/frontend-webui/tests/spec/keyboard-shortcuts.spec.js to
  its pre-gh#27080 state (commit b25ee8e0c28).

All 14 tests pass locally against mfstack 5.176-inner-silence-hotfix
with the patched frontend bundle. No change to the fix itself.
